feat: refactor admin area
