+++ src/app/(auth)/admin/backoffice/page.tsx
@@ -3,7 +3,7 @@ import CreateProfileForm from '@/components/forms/create-profile';
 import { Card, CardHeader, CardTitle } from '@/components/ui/card';
 import { City, State } from '@/types';
 
-export async function Professionals() {
+export async function Customers() {
   const states: State[] = await getStates();
   const cities: City[] = await getAllCities();
 
@@ -11,17 +11,17 @@ export async function Professionals() {
     <div className='grid gap-6'>
       <Card id='novo-perfil'>
         <CardHeader className='flex justify-between'>
-          <CardTitle> Cadastro de Perfil</CardTitle>
+          <CardTitle>Cadastro de Operador do Backoffice</CardTitle>
         </CardHeader>
 
         <CreateProfileForm
           states={states}
           cities={cities}
-          profile='PROFESSIONAL'
+          profile='BACKOFFICE'
         />
       </Card>
     </div>
   );
 }
 
-export default Professionals;
+export default Customers;
