+++ src/app/(auth)/admin/components/sidebar.tsx
@@ -8,7 +8,7 @@ import {
   UserPlusIcon,
   UserRoundCheck,
 } from '@/components/icons';
-import { LayoutDashboard } from 'lucide-react';
+import { LayoutDashboard, UserCog2, UserPlus2 } from 'lucide-react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 
@@ -27,6 +27,7 @@ function Sidebar() {
           <LayoutDashboard className='h-4 w-4' />
           Visão Geral
         </Link>
+
         <div>
           <h2 className='text-sm font-semibold text-gray-500 uppercase tracking-wide dark:text-gray-400'>
             Profissionais
@@ -83,6 +84,44 @@ function Sidebar() {
             >
               <ListIcon className='h-4 w-4' />
               Todas as vagas
+            </Link>
+          </div>
+        </div>
+
+        <div>
+          <h2 className='text-sm font-semibold text-gray-500 uppercase tracking-wide dark:text-gray-400'>
+            Clientes
+          </h2>
+          <div className='mt-3 space-y-2'>
+            <Link
+              className={`flex items-center gap-2 p-2 rounded text-sm font-medium text-gray-500 hover:text-gray-900 transition-colors dark:text-gray-300 dark:hover:text-gray-100 ${
+                pathname === '/admin/customers'
+                  ? 'bg-brand-400/20 text-gray-900'
+                  : ''
+              }`}
+              href='/admin/customers'
+            >
+              <UserPlus2 className='h-4 w-4' />
+              Novo Cliente
+            </Link>
+          </div>
+        </div>
+
+        <div>
+          <h2 className='text-sm font-semibold text-gray-500 uppercase tracking-wide dark:text-gray-400'>
+            Backoffice
+          </h2>
+          <div className='mt-3 space-y-2'>
+            <Link
+              className={`flex items-center gap-2 p-2 rounded text-sm font-medium text-gray-500 hover:text-gray-900 transition-colors dark:text-gray-300 dark:hover:text-gray-100 ${
+                pathname === '/admin/backoffice'
+                  ? 'bg-brand-400/20 text-gray-900'
+                  : ''
+              }`}
+              href='/admin/backoffice'
+            >
+              <UserCog2 className='h-4 w-4' />
+              Novo Operador
             </Link>
           </div>
         </div>
